+++ src/actions/create_payment.ts
@@ -37,6 +37,23 @@ export const create_payment = async (event: EVENT, options: Config) => {
     } catch (e) {}
 
     const repzo_serial_number = body?.serial_number?.formatted;
+    try {
+      if (body?._id) {
+        body.integration_meta = body?.integration_meta || {};
+        body.integration_meta.sync_to_sap_started = true;
+        body.integration_meta.sync_to_sap_succeeded = false;
+        await repzo.payment.update(body._id, {
+          integration_meta: body.integration_meta,
+        });
+      }
+    } catch (e) {
+      console.error(e);
+      await actionLog
+        .addDetail(
+          `Failed updating integration_meta of Payment: ${repzo_serial_number}`
+        )
+        .commit();
+    }
 
     await actionLog
       .addDetail(`Payment - ${repzo_serial_number} => ${body?.sync_id}`)
@@ -122,6 +139,23 @@ export const create_payment = async (event: EVENT, options: Config) => {
     const result = await _create(SAP_HOST_URL, "/AddPayment", sap_payment);
 
     // console.log(result);
+
+    try {
+      if (body?._id) {
+        body.integration_meta = body?.integration_meta || {};
+        body.integration_meta.sync_to_sap_succeeded = true;
+        await repzo.payment.update(body._id, {
+          integration_meta: body.integration_meta,
+        });
+      }
+    } catch (e) {
+      console.error(e);
+      await actionLog
+        .addDetail(
+          `Failed updating integration_meta of Payment: ${repzo_serial_number}`
+        )
+        .commit();
+    }
 
     await actionLog
       .addDetail(`SAP Responded with `, result)